tidying up

+++ scripts/merge_nc_trajectories.sh
@@ -15,11 +15,12 @@ mod_files=()
 DIR=$1
 output_file="$DIR/_dataset.nc"
 
-rm "$output_file"
+if [ -f "$output_file" ]; then
+    rm "$output_file"
+fi
 
 # Step 1: Add 'trajectory' dimension to each file
 for file_in in "$DIR"/*.nc; do
-    echo $file_in
     # skip file beginning with '_'
     if [[ $(basename "$file_in") == _* ]]; then
         echo "Skipping file: $file_in"
@@ -34,8 +35,8 @@ echo "Done appending dims"
 # Step 2: Concatenate the modified files along the 'trajectory' dimension
 output_file="$DIR/_dataset.nc"
 ncrcat ${ncrcat_options} "${mod_files[@]}" "${output_file}"
-ncatted -a history,global,d,, ${output_file} ${output_file} -O
-# Optional: Clean up temporary files
+ncatted -a history,global,d,, "${output_file}" -O
+# # Optional: Clean up temporary files
 rm "${mod_files[@]}"
 
 echo "Merged file created: ${output_file}"
